refactor: rename  API
