+++ packages/iceworks-server/src/app/io/controller/home/help.ts
@@ -1,4 +1,3 @@
-
 import * as path from 'path';
 
 export default (app) => {
