+++ packages/iceworks-server/src/app/io/controller/project/index.ts
@@ -1,9 +1,7 @@
 
 import * as openFolder from 'open';
 import * as openEditor from 'open-editor';
-import * as path from 'path';
 import storage from '../../../../lib/storage';
-import scanDirectory from '../../../../lib/scanDirectory';
 
 export default (app) => {
   const { Controller } = app;
@@ -46,30 +44,6 @@ export default (app) => {
       const { path } = args;
 
       return await projectManager.setCurrent(path);
-    }
-
-    async getWorkFolder() {
-      const workFolder = storage.get('workFolder');
-      const directories = await scanDirectory(workFolder);
-      return {
-        path: workFolder,
-        directories,
-      };
-    }
-
-    async setWorkFolder(ctx) {
-      const { args } = ctx;
-      const { path: setPath } = args;
-
-      const workFolder = storage.get('workFolder');
-      const newWorkFolder = path.join(workFolder, setPath);
-      storage.set('workFolder', newWorkFolder);
-
-      const directories = await scanDirectory(newWorkFolder);
-      return {
-        path: newWorkFolder,
-        directories,
-      };
     }
 
     async openFolder(ctx) {
